upd: add role business logic basic

+++ internal/usecase/human_resource_management/role/usecase/role.usecase.go
@@ -2,29 +2,61 @@ package role_usecase
 
 import (
 	"context"
+	"errors"
 	"github.com/allegro/bigcache/v3"
 	"go.mongodb.org/mongo-driver/bson/primitive"
+	"log"
+	employees_domain "shop_erp_mono/internal/domain/human_resource_management/employees"
 	roledomain "shop_erp_mono/internal/domain/human_resource_management/role"
+	userdomain "shop_erp_mono/internal/domain/human_resource_management/user"
 	"shop_erp_mono/internal/usecase/human_resource_management/role/validate"
+	"shop_erp_mono/pkg/shared/constant"
 	"time"
 )
 
 type roleUseCase struct {
-	contextTimeout time.Duration
-	roleRepository roledomain.IRoleRepository
-	cache          *bigcache.BigCache
+	contextTimeout     time.Duration
+	roleRepository     roledomain.IRoleRepository
+	userRepository     userdomain.IUserRepository
+	employeeRepository employees_domain.IEmployeeRepository
+	cache              *bigcache.BigCache
 }
 
-func NewRoleUseCase(contextTimeout time.Duration, roleRepository roledomain.IRoleRepository,
-	cacheTTL time.Duration) roledomain.IRoleUseCase {
+func NewRoleUseCase(contextTimeout time.Duration, roleRepository roledomain.IRoleRepository, userRepository userdomain.IUserRepository,
+	employeeRepository employees_domain.IEmployeeRepository, cacheTTL time.Duration) roledomain.IRoleUseCase {
 	cache, err := bigcache.New(context.Background(), bigcache.DefaultConfig(cacheTTL))
 	if err != nil {
 		return nil
 	}
-	return &roleUseCase{contextTimeout: contextTimeout, cache: cache, roleRepository: roleRepository}
+	return &roleUseCase{contextTimeout: contextTimeout, cache: cache, roleRepository: roleRepository,
+		employeeRepository: employeeRepository, userRepository: userRepository}
 }
 
-func (r *roleUseCase) CreateOne(ctx context.Context, input *roledomain.Input) error {
+func (r *roleUseCase) checkRoleLevelFromUserID(ctx context.Context, idUser string) (int, error) {
+	userID, err := primitive.ObjectIDFromHex(idUser)
+	if err != nil {
+		log.Printf("error of convert id to hex %s", err)
+	}
+
+	userData, err := r.userRepository.GetByID(ctx, userID)
+	if err != nil {
+		return 0, err
+	}
+
+	employeeData, err := r.employeeRepository.GetByID(ctx, userData.EmployeeID)
+	if err != nil {
+		return 0, err
+	}
+
+	roleLevel, err := r.roleRepository.GetByID(ctx, employeeData.RoleID)
+	if err != nil {
+		return 0, err
+	}
+
+	return roleLevel.Level, nil
+}
+
+func (r *roleUseCase) CreateOne(ctx context.Context, input *roledomain.Input, idUser string) error {
 	ctx, cancel := context.WithTimeout(ctx, r.contextTimeout)
 	defer cancel()
 
@@ -34,7 +66,7 @@ func (r *roleUseCase) CreateOne(ctx context.Context, input *roledomain.Input) er
 
 	role := &roledomain.Role{
 		ID:          primitive.NewObjectID(),
-		Title:       input.Title,
+		Name:        input.Name,
 		Description: input.Description,
 		CreatedAt:   time.Now(),
 		UpdatedAt:   time.Now(),
@@ -43,11 +75,11 @@ func (r *roleUseCase) CreateOne(ctx context.Context, input *roledomain.Input) er
 	return r.roleRepository.CreateOne(ctx, role)
 }
 
-func (r *roleUseCase) GetByTitle(ctx context.Context, title string) (roledomain.Output, error) {
+func (r *roleUseCase) GetByName(ctx context.Context, name string) (roledomain.Output, error) {
 	ctx, cancel := context.WithTimeout(ctx, r.contextTimeout)
 	defer cancel()
 
-	roleData, err := r.roleRepository.GetByTitle(ctx, title)
+	roleData, err := r.roleRepository.GetByName(ctx, name)
 	if err != nil {
 		return roledomain.Output{}, err
 	}
@@ -79,6 +111,50 @@ func (r *roleUseCase) GetByID(ctx context.Context, id string) (roledomain.Output
 	return output, nil
 }
 
+func (r *roleUseCase) GetByEnable(ctx context.Context, enable int) ([]roledomain.Output, error) {
+	ctx, cancel := context.WithTimeout(ctx, r.contextTimeout)
+	defer cancel()
+
+	roleData, err := r.roleRepository.GetByEnable(ctx, enable)
+	if err != nil {
+		return nil, err
+	}
+
+	var outputs []roledomain.Output
+	outputs = make([]roledomain.Output, 0, len(roleData))
+	for _, role := range roleData {
+		output := roledomain.Output{
+			Role: role,
+		}
+
+		outputs = append(outputs, output)
+	}
+
+	return outputs, nil
+}
+
+func (r *roleUseCase) GetByLevel(ctx context.Context, level int) ([]roledomain.Output, error) {
+	ctx, cancel := context.WithTimeout(ctx, r.contextTimeout)
+	defer cancel()
+
+	roleData, err := r.roleRepository.GetByLevel(ctx, level)
+	if err != nil {
+		return nil, err
+	}
+
+	var outputs []roledomain.Output
+	outputs = make([]roledomain.Output, 0, len(roleData))
+	for _, role := range roleData {
+		output := roledomain.Output{
+			Role: role,
+		}
+
+		outputs = append(outputs, output)
+	}
+
+	return outputs, nil
+}
+
 func (r *roleUseCase) GetAll(ctx context.Context) ([]roledomain.Output, error) {
 	ctx, cancel := context.WithTimeout(ctx, r.contextTimeout)
 	defer cancel()
@@ -101,7 +177,7 @@ func (r *roleUseCase) GetAll(ctx context.Context) ([]roledomain.Output, error) {
 	return outputs, nil
 }
 
-func (r *roleUseCase) UpdateOne(ctx context.Context, id string, input *roledomain.Input) error {
+func (r *roleUseCase) UpdateOne(ctx context.Context, id string, input *roledomain.Input, idUser string) error {
 	ctx, cancel := context.WithTimeout(ctx, r.contextTimeout)
 	defer cancel()
 
@@ -116,8 +192,9 @@ func (r *roleUseCase) UpdateOne(ctx context.Context, id string, input *roledomai
 
 	role := &roledomain.Role{
 		ID:          roleID,
-		Title:       input.Title,
+		Name:        input.Name,
 		Description: input.Description,
+		Level:       input.Level,
 		CreatedAt:   time.Now(),
 		UpdatedAt:   time.Now(),
 	}
@@ -125,7 +202,7 @@ func (r *roleUseCase) UpdateOne(ctx context.Context, id string, input *roledomai
 	return r.roleRepository.UpdateOne(ctx, roleID, role)
 }
 
-func (r *roleUseCase) DeleteOne(ctx context.Context, id string) error {
+func (r *roleUseCase) DeleteOne(ctx context.Context, id string, idUser string) error {
 	ctx, cancel := context.WithTimeout(ctx, r.contextTimeout)
 	defer cancel()
 
@@ -134,7 +211,35 @@ func (r *roleUseCase) DeleteOne(ctx context.Context, id string) error {
 		return err
 	}
 
+	count, err := r.employeeRepository.CountEmployeeByRoleID(ctx, roleID)
+	if err != nil {
+		return err
+	}
+	if count > 0 {
+		return errors.New(constant.MsgDataDeletionFailure)
+	}
+
 	return r.roleRepository.DeleteOne(ctx, roleID)
+}
+
+func (r *roleUseCase) DeleteSoft(ctx context.Context, id string, idUser string) error {
+	ctx, cancel := context.WithTimeout(ctx, r.contextTimeout)
+	defer cancel()
+
+	roleID, err := primitive.ObjectIDFromHex(id)
+	if err != nil {
+		return err
+	}
+
+	count, err := r.employeeRepository.CountEmployeeByRoleID(ctx, roleID)
+	if err != nil {
+		return err
+	}
+	if count > 0 {
+		return errors.New(constant.MsgDataDeletionFailure)
+	}
+
+	return r.roleRepository.DeleteSoft(ctx, roleID)
 }
 
 func (r *roleUseCase) CountRole(ctx context.Context) (int64, error) {
@@ -142,4 +247,9 @@ func (r *roleUseCase) CountRole(ctx context.Context) (int64, error) {
 	defer cancel()
 
 	return r.roleRepository.CountRole(ctx)
+}
+
+func (r *roleUseCase) Lifecycle(ctx context.Context) error {
+	//TODO implement me
+	panic("implement me")
 }
